API Reference

Index

SingleMachineScheduling.Instance1_rj_sumCj

SingleMachineScheduling.aprtf

SingleMachineScheduling.build_and_solve_instance

SingleMachineScheduling.build_instance_1_rj_sumCj

SingleMachineScheduling.build_solve_and_encode_instance

SingleMachineScheduling.embedding_to_sequence

SingleMachineScheduling.encoder_1_rj_sumCj

SingleMachineScheduling.evaluate_solution_1_rj_sumCj

SingleMachineScheduling.fast_local_descent_1_rj_sumCj

SingleMachineScheduling.fast_local_descent_1_rj_sumCj!

SingleMachineScheduling.milp_solve_1_rj_sumCj

SingleMachineScheduling.nb_features_encoder

SingleMachineScheduling.prtf

SingleMachineScheduling.rdi

SingleMachineScheduling.rdi!

SingleMachineScheduling.rdi_aptrf

SingleMachineScheduling.remove!

SingleMachineScheduling.sequence_to_embedding

SingleMachineScheduling.solver_name

SingleMachineScheduling.srpt_1_rj_sumCj

Docstrings

SingleMachineScheduling.Instance1_rj_sumCj — Type

struct Instance1_rj_sumCj{T} 
    nb_jobs::Int
    release_times::Vector{T}
    processing_times::Vector{T}
end

SingleMachineScheduling.aprtf — Method

function aprtf(inst::Instance1_rj_sumCj{T}) where {T}

Implements aprtf as in Chu, 1992, Efficient heuristics to minimize total flowtime with release dates. See this reference for the notations and the details of the algorithm

returns obj_value, sol

SingleMachineScheduling.build_and_solve_instance — Method

build_and_solve_instance(;seed=0,nb_jobs=50,range=0.8, solver=milp_solver, load_and_save=true)

Function that

- builds an instance `inst`
- solve the instance with `solver` to compute a solution and its value `val`
- returns `inst,val`

If load_and_save is set to true, the solution is loaded from the disk (if the solution file already exists), or computed and saved to the disk (if the solution file doesn't exit)

SingleMachineScheduling.build_instance_1_rj_sumCj — Method

function build_instance_1_rj_sumCj(;seed::Int=0,nb_jobs::Int=10,range::Float64=0.8)

generates an Instance1_rj_sumCj{T} instance with nb_jobs jobs

seed is an Integer

range is a float between 0.2 and 3 typically. Most difficult instances are obtained with range around 0.8

SingleMachineScheduling.build_solve_and_encode_instance — Method

function build_solve_and_encode_instance(;seed=0,nb_jobs=50,range=0.8, solver=milp_solver, load_and_save=true)

Function that

- builds an instance `inst`
- compute its embedding `x`
- solve the instance with `solver` to compute a solution
- computes the embedding `y` of the solution computed by `solver`
- computes the value `val` of the solution
- returns `x,y,inst,val`

If load_and_save is set to true, the solution is loaded from the disk (if the solution file already exists), or computed and saved to the disk (if the solution file doesn't exit)

SingleMachineScheduling.embedding_to_sequence — Method

function embedding_to_sequence(y::AbstractVector)

SingleMachineScheduling.encoder_1_rj_sumCj — Method

function encoder_1_rj_sumCj(inst::Instance1_rj_sumCj{T}) where {T}

returns a 27 * inst.nb_jobs matrix with the value of the features

SingleMachineScheduling.evaluate_solution_1_rj_sumCj — Method

function evaluate_solution_1_rj_sumCj(inst::Instance1_rj_sumCj,sol::Vector{I}) where {I <: Integer}

returns the objective value for the solution (permutation of the indices) encoded in sol

SingleMachineScheduling.fast_local_descent_1_rj_sumCj! — Method

function fastlocaldescent1rjsumCj!(inst::Instance1rj_sumCj,sol::Vector{I}) where {I}

Performs a fast local descent on the solution sol of instance inst. Solution sol is modified and contains the result at the end.

SingleMachineScheduling.fast_local_descent_1_rj_sumCj — Method

function fastlocaldescent1rjsumCj(inst::Instance1rj_sumCj,sol::Vector{I}) where {I}

Decoder version of fast_local_descent_1_rj_sumCj!

SingleMachineScheduling.milp_solve_1_rj_sumCj — Method

function milp_solve_1_rj_sumCj(
    inst::Instance1_rj_sumCj{T}; 
    srpt_cuts=true, 
    MILP_solver=GLPK.Optimizer
) where {T}

returns objective_value(model), solution where solution is a permutation enoced as a Vector{Int}

SingleMachineScheduling.nb_features_encoder — Method

function nb_features_encoder(encoder)

returns the number of features used by the encoder

SingleMachineScheduling.prtf — Method

prtf(inst::Instance1_rj_sumCj{T}, current_time::T, job_index) where {T}

Priority Rule for Total Flowtime (PRTF)

SingleMachineScheduling.rdi! — Method

function rdi!(
	inst::Instance1_rj_sumCj{T}, 
	dispatching_rule::AbstractVector, 
	sol::AbstractVector
) where {T}

Release Date Iteration heuristic

SingleMachineScheduling.rdi — Method

function rdi(
	inst::Instance1_rj_sumCj{T}, 
	sol::AbstractVector
) where {T}

Decoder version of rdi. Sol is used both as starting point and as guiding heuristic for rdi

SingleMachineScheduling.rdi_aptrf — Method

function rdi_aptrf(inst::Instance1_rj_sumCj)

runs rdi on inst with aptrf dispatching rule

SingleMachineScheduling.remove! — Method

function remove!(a, item)

removes all occurences of item in a

SingleMachineScheduling.sequence_to_embedding — Method

function sequence_to_embedding(seq::Vector{I}) where {I <: Int}

SingleMachineScheduling.solver_name — Method

solver_name(solver)

returns the name of solver used for logging purpose

SingleMachineScheduling.srpt_1_rj_sumCj — Method

function srpt_1_rj_sumCj(inst::Instance1_rj_sumCj{T}) where {T}

computes the preemptive solution of inst, and returns

struct Preemptive_solution_1_rj_sumCj{T}
    obj_value::T
    job_sequence::Vector{Int}
    processing_time_sequence::Vector{T}
    completion_times::Vector{T}
end

where obj_value is the preemptive objective value